% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekbohm-mle-test.R
\name{ekbohm.mle.test}
\alias{ekbohm.mle.test}
\title{Ekbohm’s MLE-based test under homoscedasticity}
\usage{
ekbohm.mle.test(x, y, alternative = c("two-sided", "greater", "less"))
}
\arguments{
\item{x}{a non-empty numeric vector of data values}

\item{y}{a non-empty numeric vector of data values}

\item{alternative}{specification of the alternative hypothesis.
Takes values: "two.sided", "greater", or "less".}
}
\value{
p-value corresponding with the hypothesis test
}
\description{
\code{ekbohm.mle.test} uses Ekbohm’s MLE-based test under homoscedasticity
to obtain a p-value for a partially matched pairs test.
}
\details{
These are the details
}
\examples{
This is an example.

}
\references{
Kuan, Pei Fen, and Bo Huang. "A simple and robust method for partially
matched samples using the p‐values pooling approach." Statistics in 
medicine 32.19 (2013): 3247-3259.
}
