% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liptak-ztest.R
\name{liptak.ztest}
\alias{liptak.ztest}
\title{Liptak weighted Z-test}
\usage{
liptak.ztest(x, y, alternative = c("two.sided", "greater", "less"))
}
\arguments{
\item{x}{a non-empty numeric vector containing some NA values}

\item{y}{a non-empty numeric vector containing some NA values}

\item{alternative}{specification of the alternative hypothesis.
Takes values: \code{"two.sided"}, \code{"greater"}, or \code{"less"}.}
}
\value{
p-value associated with the hypothesis test
}
\description{
\code{liptak.ztest} returns the p-value associated with Liptak's weighted
z-test for partially matched pairs
}
\details{
Liptak's weighted Z-test computes the p-values of a paired sample t-test on
the n1 paired entries and of a two-sample t-test on the n2+n3 unpaired
entries (see vignette for further details on the paired vs. unpaired
distinction). The two p-values are then weighted and combined as detailed in
[Kuan & Huang, 2013].

If proper sample size conditions are not met, then \code{liptak.ztest} may
exit or perform a paired or unpaired two-sample t.test, depending on the
nature of the sample size issue.

If the variance of input data is close to zero, \code{liptak.ztest} will
return an error message.
}
\examples{
In the following, the true means are not equal:

x = rnorm(400, 0, 1)
x[sample(1:400, size=75, replace=FALSE)] = NA
y = rnorm(400, 0.4, 3)
y[sample(1:400, size=75, replace=FALSE)] = NA
liptak.ztest(x, y, alternative = 'two.sided')

}
\references{
Kuan, Pei Fen, and Bo Huang. "A simple and robust method for partially
matched samples using the p‐values pooling approach." Statistics in 
medicine 32.19 (2013): 3247-3259.
}
