% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified-t-stat.R
\name{modified.t.stat}
\alias{modified.t.stat}
\title{Kim et al.'s modified t-statistic}
\usage{
modified.t.stat(x, y, alternative = c("two-sided", "greater", "less"))
}
\arguments{
\item{x}{a non-empty numeric vector of data values}

\item{y}{a non-empty numeric vector of data values}

\item{alternative}{specification of the alternative hypothesis.
Takes values: "two.sided", "greater", or "less".}
}
\value{
p-value corresponding with the hypothesis test
}
\description{
Uses Kim et al.'s modified t-statistic to obtain a p-value for
partially matched pairs.
}
\examples{


}
