% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified-t-stat.R
\name{modified.tstat}
\alias{modified.tstat}
\title{Kim et al.'s modified t-statistic}
\usage{
modified.tstat(x, y, alternative = c("two.sided", "greater", "less"))
}
\arguments{
\item{x}{a non-empty numeric vector of data values}

\item{y}{a non-empty numeric vector of data values}

\item{alternative}{specification of the alternative hypothesis.
Takes values: \code{two.sided}, \code{greater}, or \code{less}.}
}
\value{
p-value associated with the hypothesis test
}
\description{
\code{modified.tstat} uses Kim et al.'s modified t-statistic to obtain a
p-value for a partially matched pairs test.
}
\details{
Kim et al.’s modified t-statistic follows an approximately standard Gaussian
distribution under the null hypothesis. Mathematical details are provided in
[Kuan & Huang, 2013].

If proper sample size conditions are not met, then \code{modified.tstat} may
exit or perform a paired or unpaired two-sample t.test, depending on the
nature of the sample size issue.

If the variance of input data is close to zero, \code{modified.tstat} will
return an error message.
}
\examples{
In the following, the true means are not equal:

x = rnorm(400, 0, 1)
x[sample(1:400, size=75, replace=FALSE)] = NA
y = rnorm(400, 0.4, 3)
y[sample(1:400, size=75, replace=FALSE)] = NA
modified.tstat(x, y, alternative = 'two.sided')

}
\references{
Kuan, Pei Fen, and Bo Huang. "A simple and robust method for partially
matched samples using the p‐values pooling approach." Statistics in 
medicine 32.19 (2013): 3247-3259.
}
