% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-mle-test.R
\name{lin.mle.test}
\alias{lin.mle.test}
\title{Lin and Stivers’s MLE-based test under heteroscedasticity}
\usage{
lin.mle.test(x, y, alternative = c("two.sided", "greater", "less"))
}
\arguments{
\item{x}{a non-empty numeric vector containing some NA values}

\item{y}{a non-empty numeric vector containing some NA values}

\item{alternative}{specification of the alternative hypothesis.
Takes values: \code{"two.sided"}, \code{"greater"}, or \code{"less"}.}
}
\value{
p-value associated with the hypothesis test
}
\description{
\code{lin.mle.test} uses the Lin and Stivers’s MLE-based test 
under heteroscedasticity to obtain a p-value for a partially matched
pairs test.
}
\details{
Lin and Stivers’s test makes use of a modified maximum likelihood estimator
and assumption of heteroscedasticity. Under the null hypothesis, the
resulting test statistic Z_LS, follows an approximate t distribution with n1
degrees of freedom. Mathematical details are provided in [Kuan & Huang,
2013].

If proper sample size conditions are not met, then \code{lin.mle.test} may
exit or perform a paired or unpaired two-sample t.test, depending on the
nature of the sample size issue.

If the variance of input data is close to zero, \code{lin.mle.test} will
return an error message.
}
\examples{
In the following, the true means are not equal:

x = rnorm(400, 0, 1)
x[sample(1:400, size=75, replace=FALSE)] = NA
y = rnorm(400, 0.4, 3)
y[sample(1:400, size=75, replace=FALSE)] = NA
lin.mle.test(x, y, alternative = 'two.sided')

}
\references{
Kuan, Pei Fen, and Bo Huang. "A simple and robust method for partially
matched samples using the p‐values pooling approach." Statistics in 
medicine 32.19 (2013): 3247-3259.
}
