% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liptak-ztest.R
\name{liptak.ztest}
\alias{liptak.ztest}
\title{Liptak weighted Z-test}
\usage{
liptak.ztest(x, y, alternative = c("two-sided", "greater", "less"))
}
\arguments{
\item{x}{a non-empty numeric vector of data values}

\item{y}{a non-empty numeric vector of data values}

\item{alternative}{specification of the alternative hypothesis; "two.sided", "greater", or "less".}
}
\value{
p-value corresponding with the hypothesis test
}
\description{
Performs Liptak's weighted z-test for partially matched pairs
}
\examples{


}
